+++ test/ch/faults_test.exs
@@ -423,13 +423,22 @@ defmodule Ch.FaultsTest do
           :ok = :gen_tcp.close(mint)
 
           spawn_link(fn ->
-            assert {:error, %Mint.TransportError{reason: :closed}} =
-                     Ch.query(
-                       conn,
-                       "insert into unknown_table(a,b) format RowBinary",
-                       stream,
-                       encode: false
-                     )
+            err =
+              catch_error(
+                Ch.run(conn, fn conn ->
+                  stream
+                  |> Stream.into(
+                    Ch.stream(conn, "insert into unknown_table(a,b) format RowBinary\n")
+                  )
+                  |> Stream.run()
+                end)
+              )
+
+            assert {err.__struct__, Exception.message(err)} in [
+                     {DBConnection.ConnectionError,
+                      "connection is closed because of an error, disconnect or timeout"},
+                     {Mint.TransportError, "socket closed"}
+                   ]
           end)
 
           # reconnect
@@ -440,15 +449,15 @@ defmodule Ch.FaultsTest do
           :ok = :gen_tcp.send(mint, intercept_packets(clickhouse))
 
           spawn_link(fn ->
-            assert {:error, %Ch.Error{code: 60, message: message}} =
-                     Ch.query(
-                       conn,
-                       "insert into unknown_table(a,b) format RowBinary",
-                       stream,
-                       encode: false
-                     )
-
-            assert message =~ ~r/UNKNOWN_TABLE/
+            assert_raise Ch.Error, ~r/UNKNOWN_TABLE/, fn ->
+              Ch.run(conn, fn conn ->
+                stream
+                |> Stream.into(
+                  Ch.stream(conn, "insert into unknown_table(a,b) format RowBinary\n")
+                )
+                |> Stream.run()
+              end)
+            end
 
             send(test, :done)
           end)
@@ -482,13 +491,22 @@ defmodule Ch.FaultsTest do
           :ok = :gen_tcp.send(mint, intercept_packets(clickhouse))
 
           spawn_link(fn ->
-            assert {:error, %Mint.TransportError{reason: :closed}} =
-                     Ch.query(
-                       conn,
-                       "insert into unknown_table(a,b) format RowBinary",
-                       stream,
-                       encode: false
-                     )
+            err =
+              catch_error(
+                Ch.run(conn, fn conn ->
+                  stream
+                  |> Stream.into(
+                    Ch.stream(conn, "insert into unknown_table(a,b) format RowBinary\n")
+                  )
+                  |> Stream.run()
+                end)
+              )
+
+            assert {err.__struct__, Exception.message(err)} in [
+                     {DBConnection.ConnectionError,
+                      "connection is closed because of an error, disconnect or timeout"},
+                     {Mint.TransportError, "socket closed"}
+                   ]
           end)
 
           # close after first packet from mint arrives
@@ -503,15 +521,15 @@ defmodule Ch.FaultsTest do
           :ok = :gen_tcp.send(mint, intercept_packets(clickhouse))
 
           spawn_link(fn ->
-            assert {:error, %Ch.Error{code: 60, message: message}} =
-                     Ch.query(
-                       conn,
-                       "insert into unknown_table(a,b) format RowBinary",
-                       stream,
-                       encode: false
-                     )
-
-            assert message =~ ~r/UNKNOWN_TABLE/
+            assert_raise Ch.Error, ~r/UNKNOWN_TABLE/, fn ->
+              Ch.run(conn, fn conn ->
+                stream
+                |> Stream.into(
+                  Ch.stream(conn, "insert into unknown_table(a,b) format RowBinary\n")
+                )
+                |> Stream.run()
+              end)
+            end
 
             send(test, :done)
           end)